feat: refactor load balancer logic and add utility functions for NAD to interface mapping
